+++ scripts/vfx/build_openfx.sh
@@ -7,7 +7,7 @@ set -ex
 mkdir openfx
 cd openfx
 
-if [ ! -f "$DOWNLOADS_DIR/ptex-${ASWF_OPENFX_VERSION}.tar.gz" ]; then
+if [[ ! -f "$DOWNLOADS_DIR/ptex-${ASWF_OPENFX_VERSION}.tar.gz" ]]; then
     curl --location "https://github.com/AcademySoftwareFoundation/openfx/archive/refs/tags/OFX_Release_${ASWF_OPENFX_VERSION}.tar.gz" -o "$DOWNLOADS_DIR/openfx-${ASWF_OPENFX_VERSION}.tar.gz"
 fi