+++ CHANGELOG.md
@@ -9,6 +9,9 @@ All notable changes to this project will be documented in this file.
   - Update some Python dependencies
   - [Enable building iv for OpenImageIO](https://github.com/AcademySoftwareFoundation/aswf-docker/issues/357) to include Qt in ci-oiio, re-release ci-oiio 2026.3 / 2025.5 / 2024.6 images
   - Update and fix [SonarCloud scanning](https://github.com/AcademySoftwareFoundation/aswf-docker/issues/331), excluding vendored Conan recipes from analysis
+  - Update Sonar Scanner CLI to 8.0.1.6346
+  - Get rid unused build scripts
+  - Start addressing warnings from SonarCloud
 
 # 2026-02-24